# docs/migration/contracts/source-openapi.json
{
  "components": {
    "schemas": {
...
      "Body_handler_expense_upload_proof_handler_expenses__eid__proofs_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_handler_expense_upload_proof_handler_expenses__eid__proofs_post",
        "type": "object"
      },
      "Body_handler_expenses_create_handler_expenses_post": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "string"
          },
          "category": {
            "default": "commission",
            "title": "Category",
            "type": "string"
          },
          "date": {
            "default": "",
            "title": "Date",
            "type": "string"
          },
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          }
        },
        "required": [
          "reference",
          "amount",
          "file"
        ],
        "title": "Body_handler_expenses_create_handler_expenses_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/company-expenses": {
      "get": {
        "operationId": "company_expenses_list_company_expenses_get",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Company Expenses List"
      },
      "post": {
        "operationId": "company_expenses_create_company_expenses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Company Expenses Create"
      }
    },
    "/company-expenses/total": {
      "get": {
        "description": "Combined view: handler payouts + company expenses = total expenditure.",
        "operationId": "company_expenses_total_company_expenses_total_get",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Company Expenses Total"
      }
    },
    "/company-expenses/{eid}": {
      "delete": {
        "operationId": "company_expenses_delete_company_expenses__eid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "eid",
            "required": true,
            "schema": {
              "title": "Eid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Company Expenses Delete"
      },
      "patch": {
        "operationId": "company_expenses_update_company_expenses__eid__patch",
        "parameters": [
          {
            "in": "path",
            "name": "eid",
            "required": true,
            "schema": {
              "title": "Eid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Company Expenses Update"
      }
    },
...
    "/handler-expenses": {
      "get": {
        "operationId": "handler_expenses_list_handler_expenses_get",
        "parameters": [
          {
            "in": "query",
            "name": "reference",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reference"
            }
          },
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Handler Expenses List"
      },
      "post": {
        "operationId": "handler_expenses_create_handler_expenses_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_handler_expenses_create_handler_expenses_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Handler Expenses Create"
      }
    },
    "/handler-expenses/summary": {
      "get": {
        "operationId": "handler_expenses_summary_handler_expenses_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "in": "query",
            "name": "reference",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reference"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Handler Expenses Summary"
      }
    },
    "/handler-expenses/{eid}": {
      "delete": {
        "operationId": "handler_expenses_delete_handler_expenses__eid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "eid",
            "required": true,
            "schema": {
              "title": "Eid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Handler Expenses Delete"
      },
      "patch": {
        "operationId": "handler_expenses_update_handler_expenses__eid__patch",
        "parameters": [
          {
            "in": "path",
            "name": "eid",
            "required": true,
            "schema": {
              "title": "Eid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Handler Expenses Update"
      }
    },
    "/handler-expenses/{eid}/proofs": {
      "post": {
        "description": "Attach a payment screenshot to a handler expense entry.",
        "operationId": "handler_expense_upload_proof_handler_expenses__eid__proofs_post",
        "parameters": [
          {
            "in": "path",
            "name": "eid",
            "required": true,
            "schema": {
              "title": "Eid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_handler_expense_upload_proof_handler_expenses__eid__proofs_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Handler Expense Upload Proof"
      }
    },
    "/handler-expenses/{eid}/proofs/{pid}": {
      "delete": {
        "description": "Remove a proof from a handler expense entry.",
        "operationId": "handler_expense_delete_proof_handler_expenses__eid__proofs__pid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "eid",
            "required": true,
            "schema": {
              "title": "Eid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Handler Expense Delete Proof"
      },
      "get": {
        "description": "Serve a stored handler expense proof image.",
        "operationId": "handler_expense_serve_proof_handler_expenses__eid__proofs__pid__get",
        "parameters": [
          {
            "in": "path",
            "name": "eid",
            "required": true,
            "schema": {
              "title": "Eid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Handler Expense Serve Proof"
      }
    },
...
    "/handler-salaries/{reference}": {
      "delete": {
        "operationId": "handler_salaries_delete_handler_salaries__reference__delete",
        "parameters": [
          {
            "in": "path",
            "name": "reference",
            "required": true,
            "schema": {
              "title": "Reference",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Handler Salaries Delete"
      }
    },
...
  }
}